pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*"]

[project]
name = "tactical-exorcist-trpg-ai"
version = "0.1.0"
description = "タクティカル祓魔師 TRPG AI — ローカル LLM を使った TRPG ゲームマスター支援システム"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.33.0",
    "httpx>=0.27",
    "selenium>=4.0",
    "playwright>=1.40",
    "chromadb>=0.4",
    "duckduckgo-search>=6.0",
    "pydantic>=2.0",
    "rank-bm25>=0.2",
    "customtkinter>=5.2",
]

[project.optional-dependencies]
browser-use = [
    "browser-use>=0.1.40",
    "langchain-openai>=1.1.14",
    "langchain-anthropic>=0.1",
    "python-dotenv>=1.0",
]
vlm-agent = [
    "mss>=9.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8",
    "pygetwindow>=0.0.9",
    "pynput>=1.7",
    "Pillow>=10",
]
vlm-agent-som = [
    "opencv-python>=4.9",
    "numpy>=1.26",
]
portrait = [
    "Pillow>=10",
    "rembg>=2.0",
]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]

# ──────────────────────────────────────────
# pytest
# ──────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--cov=core",
    "--cov-report=term-missing",
    "--cov-report=xml:coverage.xml",
]
# tkinter / Selenium が絡む GUI テストはスキップ
markers = [
    "gui: marks tests that require tkinter (deselect with '-m not gui')",
    "integration: marks tests that require a live LM Studio server",
    "browser_use: marks tests that require Browser Use + cloud API",
    "e2e: marks end-to-end tests that require full CCFolia + LLM + browser environment",
    "vlm_agent: marks tests that require the [vlm-agent] optional extras",
]

# ──────────────────────────────────────────
# ruff (Lint + Format)
# ──────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["core", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (formatter handles this)
    "E701",  # multiple statements on one line (codebase style)
    "E702",  # multiple statements on one line (semicolon)
    "E741",  # ambiguous variable name (l used in comprehensions)
    "B008",  # do not perform function calls in default arguments
]

[tool.ruff.lint.isort]
known-first-party = ["core"]

# ──────────────────────────────────────────
# mypy
# ──────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = false
no_implicit_optional = false
warn_return_any = false
warn_unused_ignores = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["core.ccfolia_connector"]
ignore_errors = true
